Create `preprocess.input_formatter.py` to summarize formatting functions to a module.
